\name{pnea-package}
\alias{pnea-package}
\docType{package}
\title{pnea}
\description{Includes functions and examples for Parametric Network Enrichment Analysis.}
\author{Mirko Signorelli}
\keyword{package}
\seealso{\code{\link{plot.pnea}}, \code{\link{pnea}}, \code{\link{print.pnea}}, \code{\link{summary.pnea}}}
