\name{yeast}
\alias{yeast}
\docType{data}
\title{
List collecting various yeast data (see 'description')
}
\description{
\code{yeast} is a list that contains:

\code{yeastnet}: network matrix representing Yeastnet-v3 (Kim et al., 2013)

\code{ynetgenes}: vector with the names of the genes appearing in \code{yeastnet}

\code{esr1}: vector containing the first of the two gene sets that constitute the "Environmental Stress Response" (ESR) reported by Gasch et al. (2012)

\code{esr2}: vector containing the second gene set of the ESR

\code{goslimproc}: list containing the gene sets of the GOslim process ontology (Ashburner et al., 200) for the buddying yeast Saccaromyces Cerevisiae (groups 'biological process' and 'other' are not included)
}
\format{
\code{yeast}: list
}
\source{
Ashburner, M., Ball, C. A., Blake, J. A., Botstein, D., Butler, H., Cherry, J. M., Davis, A. P., Dolinski, K., Dwight, S. S., Eppig, J. T., et al. (2000). Gene ontology: tool for the unification of biology. Nat. Genet., 25(1), 25-29.

Gasch, A. P., Spellman, P. T., Kao, C. M., Carmel-Harel, O., Eisen, M. B., Storz, G., Botstein, D., and Brown, P. O. (2000). Genomic expression programs in the response of yeast cells to environmental changes. Mol. Biol. Cell, 11(12), 4241-4257.

Kim, H., Shin, J., Kim, E., Kim, H., Hwang, S., Shim, J. E., and Lee, I. (2013). Yeastnet v3: a public database of data-specific and integrated functional gene networks for saccharomyces cerevisiae. Nucleic Acids Res., pages 1-6.
}
\references{
Ashburner, M., Ball, C. A., Blake, J. A., Botstein, D., Butler, H., Cherry, J. M., Davis, A. P., Dolinski, K., Dwight, S. S., Eppig, J. T., et al. (2000). Gene ontology: tool for the unification of biology. Nat. Genet., 25(1), 25-29.

Gasch, A. P., Spellman, P. T., Kao, C. M., Carmel-Harel, O., Eisen, M. B., Storz, G., Botstein, D., and Brown, P. O. (2000). Genomic expression programs in the response of yeast cells to environmental changes. Mol. Biol. Cell, 11(12), 4241-4257.

Kim, H., Shin, J., Kim, E., Kim, H., Hwang, S., Shim, J. E., and Lee, I. (2013). Yeastnet v3: a public database of data-specific and integrated functional gene networks for saccharomyces cerevisiae. Nucleic Acids Res., pages 1-6.
}
\seealso{\code{\link{pnea}}}
\keyword{datasets}
